+++ src/catalog/index.ts
@@ -1,4 +1,5 @@
 export * from "./config";
+export * from "./projects";
 export * from "./mutations";
 export * from "./rename";
 export * from "./query";